requirements.txt
# Facebook Ads API
facebook-business>=18.0.0

# Google Ads API  
google-ads>=22.1.0

# Дополнительные утилиты
requests>=2.31.0
python-dateutil>=2.8.2

# Для разработки и тестирования
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Логирование и мониторинг
structlog>=23.1.0

# Асинхронность (опционально)
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# Форматирование и линтинг (для разработки)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Документация (опционально)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
